src/requirements.txt
bottle==0.12.17
mongo==0.2.0
dnspython==1.16.0
jsonschema==3.1.1
pandas==0.25.2
nltk==3.4.5
requests==2.18.4
pymongo==3.9.0
sklearn==0.0
numpy==1.13.3
python-dotenv==0.10.3
